[monitoring/uss_qualifier] move rid mock to rid package
