+++ monitoring/uss_qualifier/rid/mock/Dockerfile
@@ -3,10 +3,10 @@ FROM python:3.8
 # `docker build` should be run from `monitoring` (the parent folder of this folder)
 RUN apt-get update && apt-get install openssl && apt-get install ca-certificates
 RUN mkdir -p /app/monitoring/monitorlib
-RUN mkdir -p /app/monitoring/rid_qualifier/mock
+RUN mkdir -p /app/monitoring/uss_qualifier/rid/mock
 COPY monitorlib/requirements.txt /app/monitoring/monitorlib/requirements.txt
-COPY rid_qualifier/mock/requirements.txt /app/monitoring/rid_qualifier/mock/requirements.txt
-WORKDIR /app/monitoring/rid_qualifier/mock
+COPY uss_qualifier/rid/mock/requirements.txt /app/monitoring/uss_qualifier/rid/mock/requirements.txt
+WORKDIR /app/monitoring/uss_qualifier/rid/mock
 RUN pip install -r requirements.txt
 RUN rm -rf __pycache__
 ADD . /app/monitoring
@@ -15,4 +15,4 @@ ARG version
 ENV CODE_VERSION=$version
 EXPOSE 5000
 
-CMD ["gunicorn", "--preload", "--workers=1", "--bind=0.0.0.0:5000", "monitoring.rid_qualifier.mock:webapp"]
+CMD ["gunicorn", "--preload", "--workers=1", "--bind=0.0.0.0:5000", "monitoring.uss_qualifier.rid.mock:webapp"]
